TST #289 keep Padraic happy
