+++ hkl/tests/test_configuration.py
@@ -15,7 +15,7 @@ from ..configuration import DCReflection
 from ..configuration import DCSample
 from ..util import Constraint
 from ..util import new_lattice
-from . import TWO_PI
+from .tools import TWO_PI
 
 TEST_CONFIG_FILE = "data/e4c-config.json"
